Fixed execution of runnign jobs when an error occurs in one of them: All currently running jobs are finished before exiting.

+++ snakemake/snakemake/__init__.py
@@ -77,7 +77,7 @@ def snakemake(snakefile, list = False, jobs = 1, directory = None, targets = Non
 			stats.writerow(("Overall runtime", s))
 		os.chdir(olddir)
 
-		if standalone and ret == 1:
+		if False and standalone and ret == 1:
 			try:
 				# make sure ill behaving child processes are really killed (this will fail if snakemake is called programatically since it will kill the whole process)
 				os.killpg(0, signal.SIGKILL)
